Refactor offers with lenders

+++ core/api/lenders/server/methods.js
@@ -1,24 +1,41 @@
 import SecurityService from '../../security';
 import LenderService from '../LenderService';
-import { lenderInsert, lenderRemove, lenderUpdate } from '../methodDefinitions';
+import {
+  lenderInsert,
+  lenderRemove,
+  lenderUpdate,
+  lenderLinkOrganisationAndContact,
+} from '../methodDefinitions';
 
 lenderInsert.setHandler((context, { lender }) => {
-  // Add security checks
-  // Example
-  // SecurityService.checkCurrentUserIsAdmin();
+  SecurityService.checkCurrentUserIsAdmin();
   return LenderService.insert(lender);
 });
 
 lenderRemove.setHandler((context, { lenderId }) => {
-  // Add security checks
-  // Example
-  // SecurityService.checkCurrentUserIsAdmin();
+  SecurityService.checkCurrentUserIsAdmin();
   return LenderService.remove(lenderId);
 });
 
 lenderUpdate.setHandler((context, { lenderId, object }) => {
-  // Add security checks
-  // Example
-  // SecurityService.checkCurrentUserIsAdmin();
+  SecurityService.checkCurrentUserIsAdmin();
   return LenderService._update({ id: lenderId, object });
 });
+
+lenderLinkOrganisationAndContact.setHandler(
+  (context, { lenderId, organisationId, contactId }) => {
+    SecurityService.checkCurrentUserIsAdmin();
+    LenderService.addLink({
+      id: lenderId,
+      linkName: 'organisation',
+      linkId: organisationId,
+    });
+    if (contactId) {
+      LenderService.addLink({
+        id: lenderId,
+        linkName: 'contact',
+        linkId: contactId,
+      });
+    }
+  },
+);
